Functions/Cloud Functions/security-question/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-firestore==2.7.2